+++ src/briefy/ws/resources/__init__.py
@@ -19,6 +19,7 @@ from pyramid.httpexceptions import HTTPNotFound as NotFound
 from pyramid.httpexceptions import HTTPUnauthorized as Unauthorized
 
 import colander
+import newrelic.agent
 import sqlalchemy as sa
 
 
@@ -39,6 +40,13 @@ class BaseResource:
         """Initialize the service."""
         self.context = context
         self.request = request
+        cls_ = self.__class__
+        self._transaction_name = '{0}:{1}'.format(cls_.__module__, cls_.__name__)
+
+    def set_transaction_name(self, suffix):
+        """Set newrelic transaction name."""
+        name = '{0}.{1}'.format(self._transaction_name, suffix)
+        newrelic.agent.set_transaction_name(name, 'WebService')
 
     @property
     def session(self):
@@ -426,6 +434,7 @@ class RESTService(BaseResource):
 
         :returns: Newly created instance
         """
+        self.set_transaction_name('collection_post')
         request = self.request
         payload = request.validated
         model = model if model else self.model
@@ -461,6 +470,7 @@ class RESTService(BaseResource):
     @view(validators='_run_validators', permission='list')
     def collection_head(self):
         """Return the header with total objects for this request."""
+        self.set_transaction_name('collection_head')
         headers = self.request.response.headers
         total_records = self.count_records()
         headers['Total-Records'] = '{total}'.format(total=total_records)
@@ -471,6 +481,7 @@ class RESTService(BaseResource):
 
         :returns: Payload with total records and list of objects
         """
+        self.set_transaction_name('collection_get')
         headers = self.request.response.headers
         pagination = self.get_records()
         total = pagination['total']
@@ -486,6 +497,7 @@ class RESTService(BaseResource):
     @view(validators='_run_validators', permission='view')
     def get(self):
         """Get an instance of the model object."""
+        self.set_transaction_name('get')
         id = self.request.matchdict.get('id', '')
         obj = self.get_one(id, permission='view')
         return obj
@@ -493,6 +505,7 @@ class RESTService(BaseResource):
     @view(validators='_run_validators', permission='edit')
     def put(self):
         """Update an existing object."""
+        self.set_transaction_name('put')
         id = self.request.matchdict.get('id', '')
         obj = self.get_one(id, permission='edit')
         try:
@@ -520,6 +533,7 @@ class RESTService(BaseResource):
     @view(permission='delete')
     def delete(self):
         """Soft delete an object if there is an appropriated transition for it."""
+        self.set_transaction_name('delete')
         obj = self.get_one(id, permission='delete')
         # We do not delete things from the Database -
         # The delete event should be enough to trigger
@@ -575,6 +589,7 @@ class WorkflowAwareResource(BaseResource):
 
         :returns: Newly created instance
         """
+        self.set_transaction_name('collection_post')
         transition = self.request.validated['transition']
         message = self.request.validated.get('message', '')
         fields = self.request.validated.get('fields', {})
@@ -617,6 +632,7 @@ class WorkflowAwareResource(BaseResource):
     @view(validators='_run_validators')
     def collection_get(self):
         """Return the list of available transitions for this user in this object."""
+        self.set_transaction_name('collection_get')
         response = {
             'transitions': [],
             'total': 0